fix: refactor crawler worker iqwiki and farca
