+++ service/crawler/internal/crawler/farcaster/farcaster.go
@@ -12,11 +12,11 @@ import (
 	"github.com/naturalselectionlabs/pregod/common/database"
 	"github.com/naturalselectionlabs/pregod/common/database/model"
 	"github.com/naturalselectionlabs/pregod/common/database/model/metadata"
+	"github.com/naturalselectionlabs/pregod/common/datasource/farcaster"
 	"github.com/naturalselectionlabs/pregod/common/protocol"
 	"github.com/naturalselectionlabs/pregod/common/protocol/filter"
 	"github.com/naturalselectionlabs/pregod/common/utils/loggerx"
 	"github.com/naturalselectionlabs/pregod/common/utils/opentelemetry"
-	"github.com/naturalselectionlabs/pregod/common/worker/farcaster"
 	"github.com/naturalselectionlabs/pregod/service/crawler/internal/crawler"
 	lop "github.com/samber/lo/parallel"
 	"go.opentelemetry.io/otel"
